+++ android/samples/westerosemployees-hybrid/src/main/java/com/liferay/mobile/screens/westerosemployees_hybrid/utils/CardState.java
@@ -1,0 +1,8 @@
+package com.liferay.mobile.screens.westerosemployees_hybrid.utils;
+
+/**
+ * @author Víctor Galán Grande
+ */
+public enum CardState {
+	MINIMIZED, NORMAL, MAXIMIZED, BACKGROUND, HIDDEN
+}