+++ android/samples/westerosemployees-hybrid/src/main/res/raw/user_profile_portlet_js.js
@@ -1,0 +1,13 @@
+function callOffice() {
+    window.plugins.CallNumber.callNumber(
+        function(result) {
+            console.log('Success:' + result);
+        },
+        function(result) {
+            console.log('Error:' + result);
+        },
+        '900900900',
+        true);
+}
+
+window.callOffice = callOffice;