+++ ios/Framework/Themes/Default/Portlet/PortletDisplayView_default.swift
@@ -15,6 +15,7 @@
 import UIKit
 import WebKit
 
+@objc(PortletDisplayView_default)
 open class PortletDisplayView_default: BaseScreenletView, PortletDisplayViewModel {
 
 	// MARK: Public properties
@@ -68,6 +69,8 @@ open class PortletDisplayView_default: BaseScreenletView, PortletDisplayViewMode
 			screensWebView = ScreensWKWebView(jsCallHandler: jsCallHandler,
 					jsErrorHandler: jsErrorHandler, onPageLoadFinished:
 					onPageLoadFinishedHandler)
+
+			(screensWebView as? ScreensWKWebView)?.viewController = screenlet?.presentingViewController
 		}
 
 		addWebView()
@@ -99,7 +102,7 @@ open class PortletDisplayView_default: BaseScreenletView, PortletDisplayViewMode
 
     open override func onCreated() {
         super.onCreated()
-
+        backgroundColor = .clear
         self.progressPresenter = createProgressPresenter()
     }
 
@@ -116,7 +119,7 @@ open class PortletDisplayView_default: BaseScreenletView, PortletDisplayViewMode
 		guard let webView = screensWebView?.view else { return }
 
 		webView.translatesAutoresizingMaskIntoConstraints = false
-
+    
 		webView.backgroundColor = .clear
 
 		addSubview(webView)